third_party/blink/web_tests/fast/forms/select/select-set-length-expected.txt

Tests that setting the .length of an HTMLSelectElement correctly creates and destroys options.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS sel.length is 0
PASS sel.length is 200
PASS sel.length is 100
PASS sel.length is 180
PASS successfullyParsed is true

TEST COMPLETE
